fix dataset file relocation connections
