+++ dsgrid/config/dimension_config.py
@@ -71,24 +71,3 @@ class DimensionConfig(DimensionBaseConfigWithFiles):
 
         """
         return [x.dict() for x in self.model.records]
-
-
-# class MetricDimensionConfig(DimensionConfig):
-#     """Provides an interface to MetricDimensionModel."""
-
-#     @staticmethod
-#     def model_class():
-#         return DimensionModel
-
-#     def get_unique_units(self):
-#         """Return the unique units in a dimension's records.
-
-#         Returns
-#         -------
-#         set
-#             set of str
-
-#         """
-#         return {x.unit for x in self.model.records}
-
-# TODO: add get func for fuel_id when metric is EnergyEndUse
